Moved logo to the theme directory and gave it its own styling file
