+++ src/theme/Logo/styles.module.css
@@ -1,0 +1,21 @@
+.logo {
+    width: 100%;
+    display: block;
+    background: var(--ifm-color-primary);
+    padding: 24px;
+    height: 160px;
+}
+
+.logo__image {
+    width: auto;
+    height: 100%;
+    margin: 0 auto;
+    display: block;
+}
+
+@media (max-width: 996px) {
+    .logo {
+        padding: 8px;
+        height: 72px;
+    }
+}
